Exercise 5,6

+++ src/main/java/com/higgsup/intern/w01/sec06/exercise5_6/FlatFee.java
@@ -8,7 +8,7 @@ public class FlatFee extends BankAccount
     }
 
     @Override
-    public double endMonthCharge() {
+    public int endMonthCharge() {
         return 10000;
     }
 }
