+++ src/main/java/com/higgsup/intern/w01/sec06/exercise5_6/Gamble.java
@@ -20,7 +20,7 @@ public class Gamble extends BankAccount
     }
 
     @Override
-    public double endMonthCharge() {
+    public int endMonthCharge() {
         return 0;
     }
 }
